+++ actix-http/CHANGES.md
@@ -1,11 +1,18 @@
 # Changes
 
 ## Unreleased - 2020-xx-xx
+### Added
+* Added more flexible `on_connect_ext` methods for on-connect handling. [#1754]
+
+### Changed
 * Upgrade `base64` to `0.13`.
 * Upgrade `pin-project` to `1.0`.
 * Deprecate `ResponseBuilder::{if_some, if_true}`. [#1760]
 
 [#1760]: https://github.com/actix/actix-web/pull/1760
+
+
+[#1754]: https://github.com/actix/actix-web/pull/1754
 
 
 ## 2.0.0 - 2020-09-11