+++ actix-http/CHANGES.md
@@ -1,6 +1,10 @@
 # Changes
 
 ## Unreleased - 2020-xx-xx
+### Fixed
+* Started dropping `transfer-encoding: chunked` and `Content-Length` for 1XX and 204 responses. [#1767]
+
+[#1767]: https://github.com/actix/actix-web/pull/1767
 
 
 ## 2.1.0 - 2020-10-30